crypto_dashboard/requirements.txt
matplotlib
numpy
requests
websocket-client
